src/features/instagram-automation/GoogleAdsConsentManagement/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.32.3
python-dotenv==1.1.0
google-ads==28.4.1
google-generativeai==0.8.5
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
gunicorn==21.2.0
python-multipart==0.0.6
PyPDF2==3.0.1
numpy==1.24.3